Week 7 Homework
Nils Hostage and Brett Hamre

1.) At a larger duty cycle, the difference in amplitude of the
resonance peaks is more pronounced, however, it only picks up the first
few peaks. The small duty cycle shows a more homogenous amplitude for
all the peaks but it was able to pick up resonance peaks for high
frequencies.
2.) For larger acquisition rates, the resonance peaks are more
pronounced and more detailed. Additionally, the large acquisition rates
are able to pick up resonance peaks at higher frequencies. For
describing the response of 2000Hz, a minimum time of 0.4 seconds was
needed to see the signal completely die off.
3.) When the gain of the microphone is low, the signal becomes
extremely noisy and susceptible to errors from outside vibrations;
however, due to the effects of signal averaging, having a high amount
of samples filters out a lot of errors and produces a truer signal for
a given level of gain.